Move auto-hiding of items marked for destruction to the core plugin
