+++ npm/src/cocooned/plugins/core.js
@@ -2,9 +2,30 @@ import { Add } from './core/triggers/add'
 import { Remove } from './core/triggers/remove'
 import { clickHandler, delegatedClickHandler } from '../events/handlers'
 
+function hideMarkedForDestruction (cocooned, items) {
+  items.forEach(item => {
+    const destroy = item.querySelector('input[type=hidden][name$="[_destroy]"]')
+    if (destroy === null) {
+      return
+    }
+    if (destroy.getAttribute('value') !== 'true') {
+      return
+    }
+
+    cocooned.hide(item)
+  })
+}
+
 const coreMixin = (Base) => class extends Base {
   start () {
     super.start()
+
+    const hideDestroyed = () => { hideMarkedForDestruction(this, this.selection.items) }
+
+    hideDestroyed()
+    this.container.ownerDocument.addEventListener('page:load', hideDestroyed)
+    this.container.ownerDocument.addEventListener('turbo:load', hideDestroyed)
+    this.container.ownerDocument.addEventListener('turbolinks:load', hideDestroyed)
 
     this.addTriggers = Array.from(this.container.ownerDocument.querySelectorAll(this.selection.selector('triggers.add')))
         .map(element => Add.create(element, this))
